Rename RedHat BDD Framework to Shepherd BDD
- Updated all references in workflows, documentation, and code to reflect the new name "Shepherd BDD".
- Changed GitHub Actions usage paths from LukenLarra/RedHat-BDD-Framework to LukenLarra/shepherd_bdd.
- Modified pre-commit configuration, README, and agent guide to align with the new branding.
- Adjusted Python package names and entry points to match the new framework name.
- Ensured compatibility with Python 3.11 and updated dependencies accordingly.
- Refactored CLI and utility functions to support the renamed framework.
