+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-# Pre-commit configuration for RedHat BDD Framework
+# Pre-commit configuration for Shepherd BDD
 # ==================================================
 # Para instalar:
 #   1. pip install pre-commit
