+++ AGENTS.md
@@ -1,17 +1,17 @@
-# AI Agent Guide for RedHat-BDD-Framework
+# AI Agent Guide for Shepherd BDD
 
 This file is for AI coding agents working in this repository.
 
 ## What this repo is
 
 - A Python-based BDD framework plus example support files.
-- Core app under `redhat_bdd_framework/`.
+- Core app under `shepherd_bdd/`.
 - Scripts in `scripts/` for report generation and test utilities.
 - Custom GitHub Actions in `actions/` for discovery, test execution, and report publishing.
 
 ## Primary tasks for AI agents
 
-- Fix or extend backend Python code in `redhat_bdd_framework/`.
+- Fix or extend backend Python code in `shepherd_bdd/`.
 - Update action metadata and scripts under `actions/`.
 - Add or improve tests under `tests/`.
 - Maintain English-only comments and user-facing text in code.
@@ -23,13 +23,13 @@ Use these commands when editing or validating changes:
 - `make install-backend` — install backend dependencies from `backend/requirements.txt`
 - `make install-tests` — install test dependencies from `tests/requirements.txt`
 - `python -m unittest tests.test_junit_report_summary` — run the JUnit summary tests
-- `python -m redhat_bdd_framework --config framework.yml` — run the framework CLI
+- `python -m shepherd_bdd --config framework.yml` — run the framework CLI
 - `python tests/run_bdd_tests.py` — execute the repository BDD tests
 
 ## Style and conventions
 
 - Use English in all code comments, docstrings, log messages, CLI help text, and user-facing strings.
-- Keep Python code compatible with `python>=3.8` as defined in `pyproject.toml`.
+- Keep Python code compatible with `python>=3.11` as defined in `pyproject.toml`.
 - Follow the `ruff` configuration in `pyproject.toml` for formatting and linting.
 - Prefer helper functions in `scripts/` to be reusable and testable.
