+++ scripts/discover_features.py
@@ -16,7 +16,7 @@ sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
 
 import yaml
 
-from redhat_bdd_framework.behave_utils import (
+from shepherd_bdd.behave_utils import (
     BEHAVE_FLAGS_WITH_VALUES,
     is_feature_target,
     iterate_tokens,
